Write code mapping key strokes to movements on the game board
During board initialization, squares now include HTML data attributes
containing information about their positions on the board—specifically,
if they are along the edge of the game board. Using these data attributes,
we are now able to ignore invalid user inputs. For example, if the current
position on the board is on the upper left square, any ArrowUp or ArrowLeft
inputs will be ignored.