Refine the functionality for selecting a difficulty
We changed the board size for the easy difficulty to 50 squares.
The logic for restarting a game was extracted from the resetButton's
event listener and placed into a function. The easyButton and hardButton
event listeners now use this function to reset the resetButton's text,
the game board, and the marquee (if necessary). The generateStartAndEnd
function (previously pickStartAndEnd) now randomly picks the start and
end squares out of all available squares on the board.