Implement a countdown timer that ends the game if it reaches zero
A setInterval method, assigned to a variable, is used to manipulate
a DOM element that displays the remaining seconds the user has to
reach the goal. A couple of conditional statements inside setInterval
determines when the counter should be counting down and when it should
stop. In resetBoard(), variables and DOM elements relating to the counter
are reset when the game restarts.