Add random controls mapping
At the start of every game, the controls are randomly mapped. For example,
instead of moving up when the ArrowUp key is pressed, the key might be mapped
to move left on the board. On easy mode, the controls are mapped once at
the beginning of the game. On hard mode, the controls are remapped every
half-second, making it much harder to reach the goal.