Replace global variables and functions with JS modules
Modules are now being used to reduce the number of variables and
functions craeted under the global scope. The modules now allow
us to encapsulate related variables and functions (methods), as
well as protect the variables from unwanted, outside changes.
For example, the $clock module tracks whether the timer is
currently running and what the current count is. It also has
available, methods that can be invoked to retrieve and/or set
values relating to these tracked variables.