Rearrange and re-style HTML elements
We made the site friendlier to smaller devices, specifically tablets.
The navbar is now fixed to the left on large screens, and fixed to the
bottom on screens smaller than 800px. While implementing this, we
re-ordered some HTML elements, which are now easier to read and makes
more sense in terms of order of appearance.