Add footer and use JS to calc list item text modals
Added a footer to the page. Also replaced the hard-coded CSS styles
that positioned the list item modals on smaller screens. Using EventListeners,
the value needed to position these modals are now calculated when
the window's width is less than 800px. Conversely, when the window
width is greater than 800px, another event listener checks if the
CSS value for these modals have been reset. If not, it resets the
values back to where it should so they displaying properly on larger
screens.

JS functions that aren't related to the game itself have been extracted
into and put inside its on js file.